chore: optimize selector components
