+++ lerobot-web-fe/src/components/Selector.module.css
@@ -1,46 +1,11 @@
-.selectLabel {
-  color: #8291A0;
-  font-weight: bold;
-}
-
-.selectTrigger {
-  padding: 6px 12px;
-  border: 1px solid #d5dce5;
-  border-radius: 6px;
-  background-color: #ffffff;
-  cursor: pointer;
-  color: #1D2123;
+.selectWrapper {
   display: flex;
   align-items: center;
-  justify-content: space-between;
-  min-width: 120px;
+  gap: 12px;
 }
 
-.selectTrigger:disabled {
-  color:#919EAF;
-  background-color: #E6EBF5;
-  cursor:not-allowed;
-}
-
-.selectContent {
-  background-color: #ffffff;
-  color: #000000;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  margin-top: 4px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
-  position: relative;
-  min-width: 120px;
-}
-
-.selectItem {
-  padding: 8px 12px;
-  cursor: pointer;
+.selectWrapper label {
+  color: #555;
   font-size: 14px;
-}
-
-.selectItem[data-highlighted] {
-  background-color: #f0f0f0;
-  outline: none;
+  font-weight: normal;
 }
