+++ lerobot-web-fe/src/components/TeleoperateControlPanel.module.css
@@ -29,18 +29,6 @@
   margin-top: 0.5rem;
 }
 
-.selectWrapper {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.selectWrapper label {
-  color: #555;
-  font-size: 14px;
-  font-weight: normal;
-}
-
 .controlButton {
   display: inline-flex;
   align-items: center;
